refactor: Decomposed logic for breadcrumbs and search suggestions to custom hooks.
