+++ src/components/Header/Breadcrumbs/Breadcrumbs.tsx
@@ -1,28 +1,13 @@
 import { FC } from 'react';
-import { useLocation, NavLink, useParams } from 'react-router-dom';
+import { NavLink } from 'react-router-dom';
 
-import { useAppSelector } from '@redux';
-import { NavigationPaths, PATHS_TO_BREADCRUMBS } from '@constants';
+import { useBreadcrumbs } from '@hooks';
+import { NavigationPaths } from '@constants';
 
 import style from './Breadcrumbs.module.scss';
 
-const NOT_FOUND = 'Not found';
-
 export const Breadcrumbs: FC = () => {
-  const productsList = useAppSelector((state) => state.products.productsList);
-  const { pathname } = useLocation();
-  const { id } = useParams();
-
-  const paths = pathname.split('/').filter((path) => !!path);
-
-  const productTitle = productsList.find(({ productId }) => productId === id)
-    ?.productTitle;
-
-  const getBreadcrumb = (path: string) => {
-    const breadcrumb = PATHS_TO_BREADCRUMBS[path as NavigationPaths];
-
-    return breadcrumb || productTitle || NOT_FOUND;
-  };
+  const { paths, getBreadcrumb } = useBreadcrumbs();
 
   return (
     <nav className={style.breadcrumbs}>
