+++ src/hooks/index.ts
@@ -3,3 +3,5 @@ export * from './useDebounce';
 export * from './useSearch';
 export * from './useProductsFilter';
 export * from './useToggle';
+export * from './useSearchSuggestions';
+export * from './useBreadcrumbs';
